requirements.txt
torch==2.11.0
matplotlib==3.10.8
numpy==2.4.4
onnx==1.21.0
onnxruntime==1.23.2
open3d==0.19.0
Pillow==12.2.0
wandb==0.26.0
